+++ LeaveGo/LeaveGo/Features/Planner/PlannerView.swift
@@ -28,58 +28,60 @@ struct PlannerView: View {
             .navigationBarTitleDisplayMode(.inline)
             .navigationDestination(for: String.self) { value in
                 if value == "newPlanner" {
-                    PlannerAreaSelectionView()
+                    AreaSelectionView()
                 }
             }
         }
     }
 }
 
-struct PlannerListView: View {
-    var plannerList: [PlannerDTO]
-    @Binding var path: NavigationPath
-    
-    let columns = [
-        GridItem(.flexible(), spacing: DesignToken.Spacing.large),
-        GridItem(.flexible())
-    ]
-    
-    var body: some View {
+extension PlannerView {
+    struct PlannerListView: View {
+        var plannerList: [PlannerDTO]
+        @Binding var path: NavigationPath
         
-        ScrollView {
-            LazyVGrid(columns: columns, spacing: DesignToken.Spacing.large) {
-                ForEach(plannerList) { planner in
-                    PlannerCardView(planner: planner)
-                }
-                
-                Button {
-                    path.append("newPlanner")
-                } label: {
-                    PlannerCardView(planner: nil)
+        let columns = [
+            GridItem(.flexible(), spacing: DesignToken.Spacing.large),
+            GridItem(.flexible())
+        ]
+        
+        var body: some View {
+            
+            ScrollView {
+                LazyVGrid(columns: columns, spacing: DesignToken.Spacing.large) {
+                    ForEach(plannerList) { planner in
+                        PlannerCardView(planner: planner)
+                    }
+                    
+                    Button {
+                        path.append("newPlanner")
+                    } label: {
+                        PlannerCardView(planner: nil)
+                    }
                 }
             }
-        }
-        .padding(.horizontal, DesignToken.Spacing.large)
-    }
-}
-
-struct PlannerPlaceholderView: View {
-    @Binding var path: NavigationPath
-    
-    var body: some View {
-        VStack {
-            Spacer()
-            
-            Text("아직 등록된 여행이 없어요")
-                .foregroundStyle(.lgLabelSecondary)
-            
-            Spacer()
-            
-            PrimaryButton(
-                title: "여행 추가하기",
-                action: { path.append("newPlanner") }
-            )
             .padding(.horizontal, DesignToken.Spacing.large)
+        }
+    }
+    
+    struct PlannerPlaceholderView: View {
+        @Binding var path: NavigationPath
+        
+        var body: some View {
+            VStack {
+                Spacer()
+                
+                Text("아직 등록된 여행이 없어요")
+                    .foregroundStyle(.lgLabelSecondary)
+                
+                Spacer()
+                
+                PrimaryButton(
+                    title: "여행 추가하기",
+                    action: { path.append("newPlanner") }
+                )
+                .padding(.horizontal, DesignToken.Spacing.large)
+            }
         }
     }
 }
